+++ config/crds/edge.kcp.io_singleplacementslices.yaml
@@ -20,12 +20,12 @@ spec:
   - name: v1alpha1
     schema:
       openAPIV3Schema:
-        description: SinglePlacementSlice is the interface between "scheduling" and
-          syncing. For a given EdgePlacement object, the scheduler figures out which
-          Locations match that EdgePlacement and selects the SyncTarget to use within
-          that Location, then puts these results into some SinglePlacementSlice API
-          objects. We use potentially multiple API objects so that no one of them
-          has to get very big.
+        description: SinglePlacementSlice holds results of resolving a "where predicate".
+          For a given EdgePlacement object, the Where Resolver figures out which Locations
+          match that EdgePlacement and selects the SyncTarget to use within that Location,
+          then puts these results into some SinglePlacementSlice API objects. We use
+          potentially multiple API objects so that no one of them has to get very
+          big.
         properties:
           apiVersion:
             description: 'APIVersion defines the versioned schema of this representation
@@ -33,7 +33,7 @@ spec:
               internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources'
             type: string
           destinations:
-            description: '`destinations` holds some of the results of scheduling.'
+            description: '`destinations` holds some of the matching locations'
             items:
               description: SinglePlacement describes one Location that matches the
                 relevant EdgePlacement.
