+++ config/exports/apiexport-edge.kcp.io.yaml
@@ -8,6 +8,6 @@ spec:
   - v230618-23cd8129.customizers.edge.kcp.io
   - v230618-23cd8129.edgeplacements.edge.kcp.io
   - v230618-23cd8129.edgesyncconfigs.edge.kcp.io
-  - v230618-23cd8129.singleplacementslices.edge.kcp.io
   - v230618-23cd8129.syncerconfigs.edge.kcp.io
+  - v230630-f011ab29.singleplacementslices.edge.kcp.io
 status: {}
