--- docs/content/Coding Milestones/PoC2023q1/kubestellar-scheduler-subs/kubestellar-scheduler-1-build-kubestellar.md
@@ -1,6 +1,0 @@
-<!--kubestellar-scheduler-1-build-kubestellar-start-->
-```shell
-make build
-export PATH=$(pwd)/bin:$PATH
-```
-<!--kubestellar-scheduler-1-build-kubestellar-end-->
