--- docs/content/Coding Milestones/PoC2023q1/kubestellar-scheduler-subs/kubestellar-scheduler-exports.md
@@ -1,6 +1,0 @@
-<!--kubestellar-scheduler-exports-start-->
-```shell
-kubectl ws root:espw
-kubectl apply -f ../kubestellar/config/exports/
-```
-<!--kubestellar-scheduler-exports-end-->
