--- docs/content/Coding Milestones/PoC2023q1/kubestellar-scheduler-subs/kubestellar-scheduler-process-start-without-cd-kubestellar.md
@@ -1,7 +1,0 @@
-<!--kubestellar-scheduler-process-start-without-cd-kubestellar-start-->
-```shell
-kubectl ws root:espw
-go run cmd/kubestellar-scheduler/main.go -v 2 &
-sleep 45
-```
-<!--kubestellar-scheduler-process-start-without-cd-kubestellar-end-->
