--- docs/content/Coding Milestones/PoC2023q1/kubestellar-scheduler-subs/kubestellar-scheduler-process-start.md
@@ -1,7 +1,0 @@
-<!--kubestellar-scheduler-process-start-start-->
-```shell
-kubectl ws root:espw
-kubestellar-scheduler -v 2 &
-sleep 45
-```
-<!--kubestellar-scheduler-process-start-end-->
