+++ docs/content/Coding Milestones/PoC2023q1/where-resolver-subs/where-resolver-1-build-kubestellar.md
@@ -1,0 +1,6 @@
+<!--where-resolver-1-build-kubestellar-start-->
+```shell
+make build
+export PATH=$(pwd)/bin:$PATH
+```
+<!--where-resolver-1-build-kubestellar-end-->
