+++ docs/content/Coding Milestones/PoC2023q1/where-resolver-subs/where-resolver-2-ws-root-and-ws-create-edge.md
@@ -1,7 +1,7 @@
-<!--kubestellar-scheduler-2-ws-root-and-ws-create-edge-start-->
+<!--where-resolver-2-ws-root-and-ws-create-edge-start-->
 Next, use the command that makes sure the Edge Service Provider Workspace (ESPW), which is `root:espw`, and the TMC provider workspace (`root:compute`) are properly set up.
 
 ```shell
 kubestellar init
 ```
-<!--kubestellar-scheduler-2-ws-root-and-ws-create-edge-end-->
+<!--where-resolver-2-ws-root-and-ws-create-edge-end-->
