+++ docs/content/Coding Milestones/PoC2023q1/where-resolver-subs/where-resolver-exports.md
@@ -1,0 +1,6 @@
+<!--where-resolver-exports-start-->
+```shell
+kubectl ws root:espw
+kubectl apply -f ../kubestellar/config/exports/
+```
+<!--where-resolver-exports-end-->
