+++ docs/content/Coding Milestones/PoC2023q1/where-resolver-subs/where-resolver-imports.md
@@ -1,4 +1,4 @@
-<!--kubestellar-scheduler-imports-start-->
+<!--where-resolver-imports-start-->
 Use the user home workspace (\~) as the workload management workspace (WMW).
 ```shell
 kubectl ws \~
@@ -8,4 +8,4 @@ Bind APIs.
 ```shell
 kubectl apply -f config/imports/
 ```
-<!--kubestellar-scheduler-imports-end-->
+<!--where-resolver-imports-end-->
