+++ docs/content/Coding Milestones/PoC2023q1/where-resolver-subs/where-resolver-process-start.md
@@ -1,0 +1,7 @@
+<!--where-resolver-process-start-start-->
+```shell
+kubectl ws root:espw
+kubestellar-where-resolver -v 2 &
+sleep 45
+```
+<!--where-resolver-process-start-end-->
