+++ docs/content/Coding Milestones/PoC2023q1/where-resolver.md
@@ -1,9 +1,9 @@
 ---
-short_name: kubestellar-scheduler
-manifest_name: 'docs/content/Coding Milestones/PoC2023q1/kubestellar-scheduler.md'
+short_name: where-resolver
+manifest_name: 'docs/content/Coding Milestones/PoC2023q1/where-resolver.md'
 pre_req_name: 'docs/content/common-subs/pre-req.md'
 ---
-[![docs-ecutable - kubestellar-scheduler]({{config.repo_url}}/actions/workflows/docs-ecutable-scheduler.yml/badge.svg?branch={{config.ks_branch}})]({{config.repo_url}}/actions/workflows/docs-ecutable-scheduler.yml)&nbsp;&nbsp;&nbsp;
+[![docs-ecutable - where-resolver]({{config.repo_url}}/actions/workflows/docs-ecutable-scheduler.yml/badge.svg?branch={{config.ks_branch}})]({{config.repo_url}}/actions/workflows/docs-ecutable-scheduler.yml)&nbsp;&nbsp;&nbsp;
 {%
    include-markdown "../../common-subs/required-packages.md"
    start="<!--required-packages-start-->"
@@ -14,52 +14,52 @@ pre_req_name: 'docs/content/common-subs/pre-req.md'
    start="<!--save-some-time-start-->"
    end="<!--save-some-time-end-->"
 %}
-## Steps to try the scheduler
+## Steps to try the Where Resolver
 
 ### Pull the kcp source code, build kcp, and start kcp
 
 At this point you should have cloned the KubeStellar repo and `cd`ed into it as directed above.
 {%
-   include-markdown "kubestellar-scheduler-subs/kubestellar-scheduler-0-pull-kcp-and-kubestellar-source-and-start-kcp.md"
-   start="<!--kubestellar-scheduler-0-pull-kcp-and-kubestellar-source-and-start-kcp-start-->"
-   end="<!--kubestellar-scheduler-0-pull-kcp-and-kubestellar-source-and-start-kcp-end-->"
+   include-markdown "where-resolver-subs/where-resolver-0-pull-kcp-and-kubestellar-source-and-start-kcp.md"
+   start="<!--where-resolver-0-pull-kcp-and-kubestellar-source-and-start-kcp-start-->"
+   end="<!--where-resolver-0-pull-kcp-and-kubestellar-source-and-start-kcp-end-->"
 %}
 
 ### Build and initialize KubeStellar
 
 First build KubeStellar and add the result to your `$PATH`.
 {%
-   include-markdown "kubestellar-scheduler-subs/kubestellar-scheduler-1-build-kubestellar.md"
-   start="<!--kubestellar-scheduler-1-build-kubestellar-start-->"
-   end="<!--kubestellar-scheduler-1-build-kubestellar-end-->"
+   include-markdown "where-resolver-subs/where-resolver-1-build-kubestellar.md"
+   start="<!--where-resolver-1-build-kubestellar-start-->"
+   end="<!--where-resolver-1-build-kubestellar-end-->"
 %}
 
 {%
-   include-markdown "kubestellar-scheduler-subs/kubestellar-scheduler-2-ws-root-and-ws-create-edge.md"
-   start="<!--kubestellar-scheduler-2-ws-root-and-ws-create-edge-start-->"
-   end="<!--kubestellar-scheduler-2-ws-root-and-ws-create-edge-end-->"
+   include-markdown "where-resolver-subs/where-resolver-2-ws-root-and-ws-create-edge.md"
+   start="<!--where-resolver-2-ws-root-and-ws-create-edge-start-->"
+   end="<!--where-resolver-2-ws-root-and-ws-create-edge-end-->"
 %}
 
 ### Create the Workload Management Workspace (WMW) and bind it to the ESPW APIs
 {%
-   include-markdown "kubestellar-scheduler-subs/kubestellar-scheduler-imports.md"
-   start="<!--kubestellar-scheduler-imports-start-->"
-   end="<!--kubestellar-scheduler-imports-end-->"
+   include-markdown "where-resolver-subs/where-resolver-imports.md"
+   start="<!--where-resolver-imports-start-->"
+   end="<!--where-resolver-imports-end-->"
 %}
 
-### Run the KubeStellar Scheduler against the ESPW
-Go to the `root:espw` workspace and run the edge scheduler.
+### Run the KubeStellar Where Resolver against the ESPW
+Go to the `root:espw` workspace and run the Where Resolver.
 {%
-   include-markdown "kubestellar-scheduler-subs/kubestellar-scheduler-process-start.md"
-   start="<!--kubestellar-scheduler-process-start-start-->"
-   end="<!--kubestellar-scheduler-process-start-end-->"
+   include-markdown "where-resolver-subs/where-resolver-process-start.md"
+   start="<!--where-resolver-process-start-start-->"
+   end="<!--where-resolver-process-start-end-->"
 %}
-The outputs from the edge scheduler should be similar to:
+The outputs from the Where Resolver should be similar to:
 ``` { .bash .no-copy }
-I0605 10:53:00.156100   29786 scheduler.go:212] "Found APIExport view" exportName="edge.kcp.io" serverURL="https://192.168.1.13:6443/services/apiexport/jxch2kyb3c1h6bac/edge.kcp.io"
-I0605 10:53:00.157874   29786 scheduler.go:212] "Found APIExport view" exportName="scheduling.kcp.io" serverURL="https://192.168.1.13:6443/services/apiexport/root/scheduling.kcp.io"
-I0605 10:53:00.159242   29786 scheduler.go:212] "Found APIExport view" exportName="workload.kcp.io" serverURL="https://192.168.1.13:6443/services/apiexport/root/workload.kcp.io"
-I0605 10:53:00.261128   29786 controller.go:201] "starting controller" controller="kubestellar-scheduler"
+I0605 10:53:00.156100   29786 main.go:212] "Found APIExport view" exportName="edge.kcp.io" serverURL="https://192.168.1.13:6443/services/apiexport/jxch2kyb3c1h6bac/edge.kcp.io"
+I0605 10:53:00.157874   29786 main.go:212] "Found APIExport view" exportName="scheduling.kcp.io" serverURL="https://192.168.1.13:6443/services/apiexport/root/scheduling.kcp.io"
+I0605 10:53:00.159242   29786 main.go:212] "Found APIExport view" exportName="workload.kcp.io" serverURL="https://192.168.1.13:6443/services/apiexport/root/workload.kcp.io"
+I0605 10:53:00.261128   29786 controller.go:201] "starting controller" controller="where-resolver"
 ```
 
 ### Create the Inventory Management Workspace (IMW) and populate it with locations and synctargets
@@ -101,7 +101,7 @@ kubectl create -f config/samples/edgeplacement_all2all.yaml
 sleep 3
 ```
 
-The scheduler maintains a SinglePlacementSlice for an EdgePlacement in the same workspace.
+The Where Resolver maintains a SinglePlacementSlice for an EdgePlacement in the same workspace.
 ```shell
 kubectl get sps all2all -oyaml
 ```
@@ -174,7 +174,7 @@ metadata:
   uid: c9c0c2fc-d721-4c73-9788-e10711bad23a
 ```
 
-Feel free to change the Locations, SyncTargets, and EdgePlacements and see how the scheduler reacts.
+Feel free to change the Locations, SyncTargets, and EdgePlacements and see how the Where Resolver reacts.
 
 Your next step is to deliver a workload to a mailbox (that represents an edge location).  Go here to take the next step... (TBD)
