+++ docs/content/Getting-Started/quickstart-subs/quickstart-1-install-and-run-kubestellar.md
@@ -14,14 +14,14 @@ Check that `KubeStellar` is running:
 First, check that controllers are running with the following command:
 
 ```shell
-ps aux | grep -e mailbox-controller -e placement-translator -e kubestellar-scheduler
+ps aux | grep -e mailbox-controller -e placement-translator -e kubestellar-where-resolver
 ```
 
 which should yield something like:
 
 ``` { .sh .no-copy }
 user     1892  0.0  0.3 747644 29628 pts/1    Sl   10:51   0:00 mailbox-controller -v=2
-user     1902  0.3  0.3 743652 27504 pts/1    Sl   10:51   0:02 kubestellar-scheduler -v 2
+user     1902  0.3  0.3 743652 27504 pts/1    Sl   10:51   0:02 kubestellar-where-resolver -v 2
 user     1912  0.3  0.5 760428 41660 pts/1    Sl   10:51   0:02 placement-translator -v=2
 ```
