+++ docs/content/common-subs/required-packages.md
@@ -51,7 +51,7 @@
         tar xzf ko.tar.gz ko
         chmod +x ./ko
         ```
-        ``` title="GO - You will need GO to compile and run kcp and the KubeStellar scheduler.  Currently kcp requires go version 1.19"
+        ``` title="GO - You will need GO to compile and run kcp and the KubeStellar components.  Currently kcp requires go version 1.19"
         curl -L "https://go.dev/dl/go1.19.5.linux-$(dpkg --print-architecture).tar.gz" -o go.tar.gz
         tar -C /usr/local -xzf go.tar.gz
         rm go.tar.gz
