+++ hack/make-release-platform-archive.sh
@@ -37,7 +37,7 @@ srcdir=$(dirname "$0")
 cd "$srcdir/.."
 
 rm -rf bin/*
-make build OS="$target_os" ARCH="$target_arch" WHAT="./cmd/kubestellar-scheduler ./cmd/mailbox-controller ./cmd/placement-translator ./cmd/kubectl-kubestellar-syncer_gen"
+make build OS="$target_os" ARCH="$target_arch" WHAT="./cmd/kubestellar-where-resolver ./cmd/mailbox-controller ./cmd/placement-translator ./cmd/kubectl-kubestellar-syncer_gen"
 echo $'#!/usr/bin/env bash\necho' ${kcpe_version@Q} > bin/kubestellar-release
 chmod a+x bin/kubestellar-release
 mkdir -p build/release
