+++ pkg/where-resolver/reconcile_on_location.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package scheduler
+package where_resolver
 
 import (
 	"context"
